requirements.txt
opencv_python==3.4.2.17
scikit_image==0.16.2
imutils==0.5.3
Django==3.0.5
mahotas==1.4.9
numpy==1.18.2
ChatterBot==0.8.7
Pillow==8.1.0
scikit_learn==0.24.0
skimage==0.0
chatterbot-corpus==1.1.2